refactor code for submit and label
